code/text-encoding-converter/requirements.txt
Flask
flask-restplus
Flask-Uploads
python-magic
requests
